requirements.txt
# Django
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Celery
celery==5.3.4
redis==5.0.1

# Database
# SQLite встроен в Python

# Code quality
black==24.1.1
ruff==0.1.9
isort==5.13.2
mypy==1.8.0
pre-commit==3.6.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Utilities
python-dotenv==1.0.0
